chore(device): rename package googlehomefactory to ghdevicefactory
